+++ Source/editor/pattern/PatternEditor.cpp
@@ -65,15 +65,18 @@ void PatternEditor::paint(juce::Graphics &g) {
     g.fillRect(unoffsDrawRegion);
 
     // Draw bars
+    // TODO - this is broken, fix!!!
+    /*
     if (processor.getTimeSigDenominator() > 0 && processor.getTimeSigDenominator() <= 32) {
-        auto beat = (pixelsPerBeat * 4) / processor.getTimeSigDenominator();
-        auto bar = beat * processor.getTimeSigNumerator();
         g.setColour(Style::BAR_SHADE_COLOUR);
-        int firstBarX = unoffsDrawRegion.getX() - unoffsDrawRegion.getX() % bar - offsetX % (bar * 2);
-        for (int i = firstBarX; i < unoffsDrawRegion.getWidth(); i += bar * 2) {
-            g.fillRect(i + bar, unoffsDrawRegion.getY(), bar, unoffsDrawRegion.getHeight());
+        int barPulses = (pattern.getTimebase() * processor.getTimeSigNumerator()) / processor.getTimeSigDenominator();
+        int startingPulse = (xToPulse(0, false) / barPulses) * barPulses;
+        int endingPulse = (xToPulse(getWidth(), false) / barPulses + 1) * barPulses;
+        for (int i = startingPulse + barPulses; i < endingPulse; i += barPulses * 2) {
+            g.fillRect(pulseToX(i), unoffsDrawRegion.getY(), pulseToX(i + barPulses), unoffsDrawRegion.getHeight());
         }
     }
+    */
 
     // Draw octave 0
     auto numInputNotes = processor.getNumInputNotes();
@@ -93,20 +96,21 @@ void PatternEditor::paint(juce::Graphics &g) {
     // Draw gridlines
     // - Horizontal
     g.setColour(Style::EDITOR_GRIDLINES_COLOUR);
-    int horizontalGridlineStart = (getHeight() / 2 - offsetY) % pixelsPerNote - pixelsPerNote / 2;
-    for (int i = horizontalGridlineStart; i < getHeight(); i += pixelsPerNote) {
-        g.fillRect(0, i, getWidth(), 2);
+    int startingNote = yToNote(getHeight());
+    int endingNote = yToNote(0);
+    for (int i = startingNote; i < endingNote; i++) {
+        g.fillRect(0, noteToY(i) - 1, getWidth(), 2);
     }
 
     // - Vertical
-    float beatDiv = (static_cast<float>(pixelsPerBeat) / static_cast<float>(state.divisor));
-    int beatN = 0;
-    for (auto i = static_cast<float>((-offsetX) % pixelsPerBeat); i < static_cast<float>(getWidth()); i += beatDiv, beatN++) {
-        if (beatN % state.divisor == 0) {
-            g.fillRect(juce::roundToInt(i), 0, 4, getHeight());
-        } else {
-            g.fillRect(juce::roundToInt(i), 0, 2, getHeight());
-        }
+    int startingPulse = xToPulse(0);
+    int endingPulse = xToPulse(getWidth());
+    int pulseInc = pattern.getTimebase() / state.divisor;
+    for (int i = startingPulse; i < endingPulse; i += pulseInc) {
+        if (i % pattern.getTimebase() == 0)
+            g.fillRect(pulseToX(i) - 2, 0, 4, getHeight());
+        else
+            g.fillRect(pulseToX(i) - 1, 0, 2, getHeight());
     }
 
     // Draw octaves