+++ CHANGELOG.md
@@ -12,6 +12,7 @@ in the [wiki](https://gitlab.com/LibreArp/LibreArp/wikis/versioning-pattern).
 * **IMPROVE** Added scroll and zoom smoothing
 * **IMPROVE** Extended pattern editor tooltips
   * They now describe the function of modifiers (`Alt`, `Ctrl`, `Shift`)
+* **FIX** The grid in the pattern editor is now drawn more precisely relative to the notes
 
 ### LibreArp 2.3